feat: extract EDR context for reuse
