+++ v-next/hardhat/src/internal/builtin-plugins/network-manager/edr/edr-provider.ts
@@ -45,6 +45,7 @@ import { hexToBytes } from "ethereum-cryptography/utils";
 import { addr } from "micro-eth-signer";
 
 import { EDR_NETWORK_REVERT_SNAPSHOT_EVENT } from "../../../constants.js";
+import { getGlobalEdrContext } from "../../../edr/context.js";
 import { DEFAULT_HD_ACCOUNTS_CONFIG_PARAMS } from "../accounts/constants.js";
 import { BaseProvider } from "../base-provider.js";
 import { getJsonRpcRequest, isFailedJsonRpcResponse } from "../json-rpc.js";
@@ -54,7 +55,6 @@ import {
   UnknownError,
 } from "../provider-errors.js";
 
-import { getGlobalEdrContext } from "./edr-context.js";
 import { createSolidityErrorWithStackTrace } from "./stack-traces/stack-trace-solidity-errors.js";
 import {
   isDebugTraceResult,
